telegram_bot/requirements.txt
aiogram>=3.15.0
httpx>=0.27.0
python-dotenv>=1.0.0
requests>=2.32.5
